Fix bug where Yabbe path wasn't found if .sl2 was in separate folder
This patch fixes a bug where Yabber wasn't found if the passed .sl2 file was not in the same folder as the main module exe. This was caused by using parent path of .sl2 as a starting point for the search, this was changed to be the parent path of the base module.