Remove dependence on base module path in argv[]
We used to derive the Yabber path from the base module path in argv[]. This was fine when compiling as CLI, it's not so great for a library since the called would have to provide the baseModulePath. This patch removes the use this path and the requirement to pass it.